size_t fixes for lib/. Changed OFF_T_FORMAT to PRIuOFF_T which is more C99-like.
--HG--
branch : HEAD

+++ src/lib/base64.c
@@ -45,7 +45,7 @@
 static const char basis_64[] =
    "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789+/";
 
-const char *base64_encode(const unsigned char *data, unsigned int size)
+const char *base64_encode(const unsigned char *data, size_t size)
 {
 	char *buffer, *p;
 	int c1, c2, c3;
